backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn==1.6.1
joblib
psycopg2-binary